Made build unittest flag unique
FetchContent just calls add_subdirectory so any flags set
for the parent program also apply to the library. This means
that if the using program also has a BUILD_TEST for its test
it builds the tests also for this library. If we just put the
name of the project in this flag then it is unique to this
library and we don't have this problem anymore.