Fixed assigning other to unitialized current Get()
The Optional assignment operators didn't check if the current
value was even set. So regardless the always just called Get()
and then assigned that with the new value. This opened the
problem that the = operator might work with unitialized memory,
which is fine for a constructor but not the operator. The new
version first checks if a value is currently stored. If there is
then the = operator is called. Otherwise the constructor is
called instead.