Fixed ExtractBits not working on larger than native int types
None of the test cases covered larger than native types.
So in another they were used and caused defined yet unexpected
behaviour with data being set to '0' like it was out of range.
Now some test cases are included that use a 64-bit integer
to extract and set to some bits.